geese_plot: reimplement in C++
It's 2021, python2 is end-of-life and not installed by default
in Debian 11 or in OpenBSD. And I won't go through my code from
decades back and try to convert it to python3. Especially since
the Python community may pull a similar stunt again at any time.

And, I'm no longer that interested in Python. It's ok for quick
hacks, but I can write better code in C++ in only a little more
time.

For plotting coordinates on maps, it turns out libgd works just as
well as PIL did -- I could have migrated a decade ago, probably.

This was a one-day port, and the remaining regressions are:

- the '-t fmt' option is broken
- the program wants to kill xv; it won't react if I quit xv myself
- the program should feed xv via a pipe, not a temporary file